+++ src/App.jsx
@@ -1,18 +1,15 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
 import OfertaESpecial from './components/OfertaESpecial'
 import NovasColecoes from './components/NovasColecoes'
 import Rodape from './components/Rodape'
-import Tech3 from './components/TECH3'
+import ProductListing from './components/ProductListing'
 
 function App() {
 
   return (
     <>
       <NovasColecoes></NovasColecoes>
-      <Tech3 />
+      <ProductListing />
       <OfertaESpecial />
       <Rodape />
     </>
