+++ src/components/ProductListing.css
@@ -1,0 +1,123 @@
+.div-products {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  background-color: #F9F8FE;
+  padding: 20px;
+}
+
+.div-products-title {
+  width: 875px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 15px;
+}
+
+.products-h2 {
+  font-family: sans-serif;
+  color: #474747;
+}
+
+.products-h3 {
+  font-family: sans-serif;
+  color: #D34D8E;
+  font-size: 14px;
+  font-weight: 400;
+}
+
+.products-h3:hover {
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.products-h3::after {
+  content: '→';
+  margin-left: 8px;
+  font-size: 1em;
+  color: #D34D8E;
+}
+
+.products-img {
+  width: 150px;
+  height: 150px;
+}
+
+.div-products-content {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 16px;
+  width: 875px;
+}
+
+.div-products-item:hover {
+  cursor: pointer;
+}
+
+.div-products-item {
+  width: 201px;
+}
+
+.div-products-item1 {
+  background-color: white;
+  padding: 12px;
+  height: 200px;
+  display: flex;
+  flex-direction: column;
+}
+
+.products-p-off {
+  font-family: sans-serif;
+  position: absolute;
+  border-radius: 30px / 30px;
+  padding: 5px;
+  padding-left: 10px;
+  background-color: #E7FF86;
+  font-size: 12px;
+  font-weight: 800;
+  color: #474747;
+  width: 85px;
+  margin-top: 0;
+}
+
+.products-p1 {
+  font-family: sans-serif;
+  color: #969696;
+  font-size: 10px;
+  font-weight: 800;
+}
+
+.products-p2 {
+  font-family: sans-serif;
+  margin: 0;
+  color: #474747;
+}
+
+.products-price {
+  display: flex;
+  flex-direction: row;
+}
+
+.products-p3 {
+  font-family: sans-serif;
+  text-decoration: line-through;
+  margin-top: 7px;
+  margin-right: 10px;
+  color: #969696;
+  font-weight: 500;
+}
+
+.products-p4 {
+  font-family: sans-serif;
+  color: #474747;
+  margin-top: 7px;
+  font-weight: 600;
+}
+
+.div-products-img-div {
+  display: flex;
+  flex-direction: column;
+  margin-top: 30px;
+}
